packages/qvac-lib-infer-nmtcpp/benchmarks/client/pyproject.toml
[project]
name = "tetherto-qvac-lib-inference-addon-mlc-marian-benchmark-client"
version = "0.0.1"
description = "Benchmark client for Marian translation addons"
authors = [
    {name = "Tether"}
]
license = {text = "Apache-2.0"}
readme = "README.md"
requires-python = "^3.12.0"
dependencies = [
    "pyyaml (>=6.0.2,<7.0.0)",
    "pydantic (>=2.11.4,<3.0.0)",
    "sacrebleu (>=2.5.1,<3.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "datasets (>=3.6.0,<4.0.0)",
    "iso639-lang (>=2.6.0,<3.0.0)",
    "sentencepiece (>=0.2.0,<0.3.0)",
    "unbabel-comet (>=2.2.6,<3.0.0)",
    "huggingface-hub[hf-xet] (>=0.32.0,<0.33.0)"
]

[tool.poetry]
packages = [{include = "src"}]

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
